# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  main.ErrorResponse:
    properties:
      error:
        example: Invalid request payload
        type: string
    type: object
...
  main.ItemData:
    type: object
  main.ItemModel:
    properties:
      collection_id:
        example: 456e7890-e89b-12d3-a456-426614174001
        type: string
      created_at:
        example: "2023-01-01T10:30:00Z"
        type: string
      data:
        $ref: '#/definitions/main.ItemData'
      id:
        example: 123e4567-e89b-12d3-a456-426614174000
        type: string
      updated_at:
        example: "2023-12-01T10:30:00Z"
        type: string
    type: object
...
  main.SuccessMessage:
    properties:
      message:
        example: Operation completed successfully
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@gorectus.com
    name: GoRectus API Support
  description: A comprehensive API for managing collections, items, fields, users,
    and roles
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: GoRectus API
  version: "1.0"
paths:
...
  /items/{collection}:
    get:
      consumes:
      - application/json
      description: Retrieve a list of all items from a specific collection
      parameters:
      - description: Collection name
        in: path
        name: collection
        required: true
        type: string
      - description: Limit the number of results
        in: query
        name: limit
        type: integer
      - description: Offset for pagination
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: List of items
          schema:
            items:
              $ref: '#/definitions/main.ItemModel'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/main.ErrorResponse'
        "404":
          description: Collection not found
          schema:
            $ref: '#/definitions/main.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/main.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get all items from a collection
      tags:
      - items
    post:
      consumes:
      - application/json
      description: Create a new item in a specific collection
      parameters:
      - description: Collection name
        in: path
        name: collection
        required: true
        type: string
      - description: Item data
        in: body
        name: item
        required: true
        schema:
          $ref: '#/definitions/main.ItemModel'
      produces:
      - application/json
      responses:
        "201":
          description: Created item
          schema:
            $ref: '#/definitions/main.ItemModel'
        "400":
          description: Invalid request payload
          schema:
            $ref: '#/definitions/main.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/main.ErrorResponse'
        "404":
          description: Collection not found
          schema:
            $ref: '#/definitions/main.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/main.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a new item
      tags:
      - items
  /items/{collection}/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a specific item from a collection by its ID
      parameters:
      - description: Collection name
        in: path
        name: collection
        required: true
        type: string
      - description: Item ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success message
          schema:
            $ref: '#/definitions/main.SuccessMessage'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/main.ErrorResponse'
        "404":
          description: Item not found
          schema:
            $ref: '#/definitions/main.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/main.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete an item
      tags:
      - items
    get:
      consumes:
      - application/json
      description: Retrieve a specific item from a collection by its ID
      parameters:
      - description: Collection name
        in: path
        name: collection
        required: true
        type: string
      - description: Item ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Item details
          schema:
            $ref: '#/definitions/main.ItemModel'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/main.ErrorResponse'
        "404":
          description: Item not found
          schema:
            $ref: '#/definitions/main.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/main.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get item by ID
      tags:
      - items
    patch:
      consumes:
      - application/json
      description: Update an existing item in a specific collection
      parameters:
      - description: Collection name
        in: path
        name: collection
        required: true
        type: string
      - description: Item ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated item data
        in: body
        name: item
        required: true
        schema:
          $ref: '#/definitions/main.ItemModel'
      produces:
      - application/json
      responses:
        "200":
          description: Updated item
          schema:
            $ref: '#/definitions/main.ItemModel'
        "400":
          description: Invalid request payload
          schema:
            $ref: '#/definitions/main.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/main.ErrorResponse'
        "404":
          description: Item not found
          schema:
            $ref: '#/definitions/main.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/main.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update an existing item
      tags:
      - items
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"